// functions/src/enum/action.enum.ts
export enum actionEnum {
    create = 'create',
    update = 'update',
    complement = 'complement',
    delete = 'delete',
    save = 'save',
    accept = 'accept',
    add = 'add',
    remove = 'remove',
    conflit = 'conflit',
    send = 'send',
    reset = 'reset',
    upload = 'upload',
    setup = 'setup',
    logout = 'logout',
    processed = 'processed',
    blocked = 'blocked',
    received = 'received',
}